Reorganize HealthPage and Chart components
