+++ client/pages/DashboardPage/HealthPage/CPUUsageChart.tsx
@@ -1,22 +1,12 @@
 import React from 'react';
 import { ResponsiveLine } from '@nivo/line';
+import type { DataObj } from './charts';
 
-export type Props = NonNullable<unknown>; // TODO: figure out better type for this
-
-interface DataPoint {
-  x: number;
-  y: number | string;
-}
-interface DataObj {
-  id: string;
-  data: DataPoint[];
-}
-
-interface MemChartProps {
+interface CPUUsageChartProps {
   chartData: DataObj[]; // Update the type of chartData according to your data structure
 }
 
-const MemChart = ({ chartData }: MemChartProps) => {
+const CPUUsageChart = ({ chartData }: CPUUsageChartProps) => {
   return (
     <div style={{ height: 420, maxWidth: '100%' }}>
       <ResponsiveLine
@@ -86,4 +76,4 @@ const MemChart = ({ chartData }: MemChartProps) => {
   );
 };
 
-export default MemChart;
+export default CPUUsageChart;
