+++ client/pages/DashboardPage/HealthPage/charts.ts
@@ -1,0 +1,9 @@
+export interface DataPoint {
+  x: number;
+  y: number | string;
+}
+
+export interface DataObj {
+  id: string;
+  data: DataPoint[];
+}
